sign_language_gui.py
import tkinter as tk
from tkinter import Label, Button, Canvas
from PIL import Image, ImageTk
import cv2
import numpy as np

# Dummy recognition logic for demonstration
# Replace with your actual recognition and skeleton drawing logic

def get_dummy_character():
    return 'R'

def get_dummy_sentence():
    return 'HELLO THE'

def get_dummy_suggestions():
    return ['HE', 'THEE', 'THEN', 'THEM']

def get_dummy_skeleton():
    # Create a blank white image and draw a green skeleton-like shape
    img = np.ones((300, 300, 3), np.uint8) * 255
    pts = np.array([[150, 50], [150, 100], [120, 200], [180, 200], [150, 100], [100, 250], [200, 250]], np.int32)
    pts = pts.reshape((-1, 1, 2))
    cv2.polylines(img, [pts], False, (0, 255, 0), 3)
    return img

class SignLanguageApp:
    def __init__(self, root):
        self.root = root
        self.root.title('Sign Language To Text Conversion')
        self.root.geometry('1000x600')
        self.root.configure(bg='white')

        # Title
        Label(root, text='Sign Language To Text Conversion', font=('Courier', 24, 'bold'), bg='white').pack(pady=10)

        # Video and skeleton frames
        frame = tk.Frame(root, bg='white')
        frame.pack()
        self.video_label = Label(frame, bg='black')
        self.video_label.grid(row=0, column=0, padx=30, pady=10)
        self.skeleton_label = Label(frame, bg='white')
        self.skeleton_label.grid(row=0, column=1, padx=30, pady=10)

        # Info labels
        self.char_label = Label(root, text='Character :', font=('Courier', 16, 'bold'), bg='white')
        self.char_label.pack(anchor='w', padx=50)
        self.sentence_label = Label(root, text='Sentence :', font=('Courier', 16, 'bold'), bg='white')
        self.sentence_label.pack(anchor='w', padx=50)
        self.suggestions_label = Label(root, text='Suggestions :', font=('Courier', 16, 'bold'), fg='red', bg='white')
        self.suggestions_label.pack(anchor='w', padx=50)
        self.suggestion_buttons = []
        self.suggestion_frame = tk.Frame(root, bg='white')
        self.suggestion_frame.pack(anchor='w', padx=180)

        # Control buttons
        self.control_frame = tk.Frame(root, bg='white')
        self.control_frame.pack(anchor='e', padx=50, pady=10)
        Button(self.control_frame, text='Clear', font=('Courier', 14), command=self.clear).pack(side='left', padx=10)
        Button(self.control_frame, text='Speak', font=('Courier', 14), command=self.speak).pack(side='left', padx=10)

        # Start video capture
        self.cap = cv2.VideoCapture(0)
        self.update_video()

    def update_video(self):
        ret, frame = self.cap.read()
        if ret:
            frame = cv2.flip(frame, 1)
            img = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
            img = Image.fromarray(img)
            img = img.resize((300, 300))
            imgtk = ImageTk.PhotoImage(image=img)
            self.video_label.imgtk = imgtk
            self.video_label.configure(image=imgtk)
        else:
            self.video_label.configure(text='Camera not found', image='')

        # Update skeleton
        skeleton_img = get_dummy_skeleton()
        skeleton_img = cv2.cvtColor(skeleton_img, cv2.COLOR_BGR2RGB)
        skeleton_img = Image.fromarray(skeleton_img)
        skeleton_img = skeleton_img.resize((300, 300))
        skeleton_imgtk = ImageTk.PhotoImage(image=skeleton_img)
        self.skeleton_label.imgtk = skeleton_imgtk
        self.skeleton_label.configure(image=skeleton_imgtk)

        # Update info
        char = get_dummy_character()
        sentence = get_dummy_sentence()
        suggestions = get_dummy_suggestions()
        self.char_label.config(text=f'Character : {char}')
        self.sentence_label.config(text=f'Sentence : {sentence}')
        self.suggestions_label.config(text='Suggestions :')
        for btn in self.suggestion_buttons:
            btn.destroy()
        self.suggestion_buttons = []
        for word in suggestions:
            btn = Button(self.suggestion_frame, text=word, font=('Courier', 14))
            btn.pack(side='left', padx=5)
            self.suggestion_buttons.append(btn)

        self.root.after(30, self.update_video)

    def clear(self):
        # Implement clear logic
        pass

    def speak(self):
        # Implement speak logic
        pass

if __name__ == '__main__':
    root = tk.Tk()
    app = SignLanguageApp(root)
    root.mainloop()
